Updated Custom Service Create Page

+++ lib/models/service_model.dart
@@ -4,7 +4,7 @@ class Help4YouServices {
   final String serviceId;
   final String serviceTitle;
   final String serviceDescription;
-  final int servicePrice;
+  final double servicePrice;
   final bool visibility;
 
   Help4YouServices({
